pubsub update
