+++ apps/festival/festival/src/app/landing/landing.component.html
@@ -1,15 +1,15 @@
 <landing-shell-page>
 
-  <shell-header>
+  <landing-header>
     <h4>Welcome to Archipel Market</h4>
     <p class="mat-display-2 offer">The ongoing<br />film market</p>
     <a mat-flat-button color="accent" routerLink="/auth/connexion">
       <span>Join now</span>
       <mat-icon svgIcon="arrow_forward"></mat-icon>
     </a>
-  </shell-header>
+  </landing-header>
 
-  <shell-content>
+  <landing-content>
     <h1 class="mat-display-1">How it works</h1>
     <mat-tab-group animationDuration="1000ms" selectedIndex="0">
       <mat-tab [label]="tabNames[0]">
@@ -34,8 +34,7 @@
         </article>
       </mat-tab>
       <mat-tab [label]="tabNames[1]">
-        <article fxLayout="row wrap" fxLayoutGap="24px" fxLayoutAlign="space-between start"
-          fxLayout.lt-lg="column">
+        <article fxLayout="row wrap" fxLayoutGap="24px" fxLayoutAlign="space-between start" fxLayout.lt-lg="column">
           <img fxHide.xs [asset]="tabTitle[1].imgAsset" alt="buyer image">
           <div>
             <h3>{{ tabTitle[1].title }}</h3>
@@ -64,88 +63,6 @@
       </a>
     </section>
 
-  </shell-content>
+  </landing-content>
 
-  <!-- Learn More -->
-
-  <shell-contact>
-    <article fxLayout="column" fxLayoutAlign="start center">
-      <h1>Want to learn more?</h1>
-      <h3>Schedule a demo with our team!</h3>
-      <p>
-        Program a personalized demonstration of the platform with a member of the Archipel team.
-      </p>
-      <p>
-        Please fill out the form and we will get back to you to choose the best time.
-      </p>
-    </article>
-    <form fxLayout="column" [formGroup]="form" (ngSubmit)="sendRequest(form)">
-      <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column" fxLayoutGap="24px">
-        <mat-form-field appearance="outline">
-          <mat-label>First name</mat-label>
-          <input matInput type="text" placeholder="First name" formControlName="firstName" required />
-          <mat-error *ngIf="form.get('firstName').hasError('required')">
-            First name is <strong>required</strong>
-          </mat-error>
-        </mat-form-field>
-        <mat-form-field appearance="outline">
-          <mat-label>Last name</mat-label>
-          <input matInput type="text" placeholder="Last name" formControlName="lastName" required />
-          <mat-error *ngIf="form.get('lastName').hasError('required')">
-            Last name is <strong>required</strong>
-          </mat-error>
-        </mat-form-field>
-      </article>
-      <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column" fxLayoutGap="24px">
-        <mat-form-field appearance="outline">
-          <mat-label>Company name</mat-label>
-          <input matInput type="text" placeholder="Company name" formControlName="companyName" required />
-          <mat-error *ngIf="form.get('companyName').hasError('required')">
-            Company name is <strong>required</strong>
-          </mat-error>
-        </mat-form-field>
-        <mat-form-field appearance="outline">
-          <mat-label>Activity</mat-label>
-          <mat-select placeholder="Activity" formControlName="role" required>
-            <mat-option *ngFor="let role of roles" [value]="role">{{ role | roleLabel }}</mat-option>
-          </mat-select>
-          <mat-error *ngIf="form.get('role').hasError('required')">
-            Activity is <strong>required</strong>
-          </mat-error>
-        </mat-form-field>
-      </article>
-      <article fxLayout="row" fxLayoutAlign="space-between center" fxLayout.lt-md="column" fxLayoutGap="24px">
-        <mat-form-field appearance="outline">
-          <mat-label>Email</mat-label>
-          <input matInput type="email" placeholder="Email" formControlName="email" required />
-          <mat-error *ngIf="form.get('email').hasError('required')">
-            Email is <strong>required</strong>
-          </mat-error>
-        </mat-form-field>
-        <mat-form-field appearance="outline">
-          <tel-input [form]="form.get('phoneNumber')" placeholder="Phone number"></tel-input>
-        </mat-form-field>
-      </article>
-      <button [disabled]="submitted" fxFlexAlign="center" mat-flat-button color="accent" type="submit">Send a demo
-        request</button>
-    </form>
-  </shell-contact>
-
-
-  <!-- Footer -->
-
-  <shell-footer>
-    <footer fxLayout="column" fxLayoutAlign="start center" fxLayout.lt-md="column" fxLayoutAlign.lt-md="center center">
-      <a routerLink="/">
-        <app-logo></app-logo>
-      </a>
-      <h3>Proudly associated with the following partners</h3>
-      <span fxLayout="row wrap" fxLayoutAlign="space-evenly start" fxLayoutGap="80px">
-        <img asset="CNC_logo.webp" type="logo" alt="cnc">
-        <img src="/assets/logo/BPIfrance.png" alt="BPI France logo" />
-        <img src="/assets/logo/creativemediaeurope.png" alt="creatiev media europe logo" />
-        <img asset="IFCIC_logo.webp" type="logo" alt="ifcic">
-      </span>
-    </footer>
-  </shell-footer>
 </landing-shell-page>
