+++ libs/landing/src/lib/shell/shell.component.scss
@@ -27,3 +27,7 @@ mat-toolbar {
     }
   }
 }
+
+form {
+  
+}
