+++ libs/landing/src/lib/shell/shell.module.ts
@@ -3,18 +3,21 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { ReactiveFormsModule } from '@angular/forms';
 
 // Components
 import {
   LandingShellComponent,
-  HeaderShellComponent,
-  ShellContentComponent,
-  ShellContactComponent,
-  ShellFooterComponent
+  LandingHeaderComponent,
+  LandingContentComponent,
+  LandingContactComponent,
+  LandingFooterComponent
 } from './shell.component';
 
 // Blockframes
 import { AppLogoModule } from '@blockframes/ui/layout/app-logo/app-logo.module';
+import { TelInputModule } from '@blockframes/ui/tel-input/tel-input.module'
+import { RolePipeModule } from '@blockframes/utils/pipes/role.pipe';
 
 // Material
 import { MatButtonModule } from '@angular/material/button';
@@ -22,33 +25,42 @@ import { MatIconModule } from '@angular/material/icon';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatCardModule } from '@angular/material/card';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatSelectModule } from '@angular/material/select';
+import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
   imports: [
     CommonModule,
     RouterModule,
+    ReactiveFormsModule,
     FlexLayoutModule,
     AppLogoModule,
+    RolePipeModule,
+    TelInputModule,
 
     // Material
     MatButtonModule,
     MatIconModule,
     MatToolbarModule,
     MatCardModule,
+    MatFormFieldModule,
+    MatSelectModule,
+    MatInputModule
   ],
   exports: [
     LandingShellComponent,
-    HeaderShellComponent,
-    ShellContentComponent,
-    ShellContactComponent,
-    ShellFooterComponent,
+    LandingHeaderComponent,
+    LandingContentComponent,
+    LandingContactComponent,
+    LandingFooterComponent,
     AppLogoModule],
   declarations: [
     LandingShellComponent,
-    HeaderShellComponent,
-    ShellContentComponent,
-    ShellContactComponent,
-    ShellFooterComponent
+    LandingHeaderComponent,
+    LandingContentComponent,
+    LandingContactComponent,
+    LandingFooterComponent
   ],
 })
 export class LandingShellModule { }
